requirements.txt
gradio==4.44.0
anthropic==0.40.0
chromadb==0.5.23
pypdf2==3.0.1
sentence-transformers==3.3.1
python-dotenv==1.0.1